KAFKA-16764: New consumer should throw InvalidTopicException on poll when invalid topic in metadata
- move integration test
